third_party/WebKit/LayoutTests/http/tests/priority-hints/image-low-importance-in-viewport.html

<script src=resources/common.js></script>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>

<script>
  let t = async_test('low importance on in-viewport images must be fetched with kLow resource load priority');
  // Synchronous wait to delay layout by .5 seconds
  let end = Date.now() + 500;
  while (Date.now() < end) {}
</script>

<!--
  In-viewport images get re-prioritized if layout is computed mid-flight. With importance=low, re-prioritization should have no effect,
  we delay both the image request (will be fetched by preload scanner) by a second, and finish layout mid-flight. As long as the final
  resource load priority is kLow, we know the image was not re-prioritized.
-->
<img importance=low src="../resources/slow-image.php?sleep=1000&name=square.png" alt=img onload="assert_priority_onload(this.src, kLow, t)()">
